.......... [ZBX-4311] merged changes from trunk
git-svn-id: svn://svn.zabbix.com/branches/dev@23745 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-4311/include/db.h
@@ -574,7 +574,7 @@ double	DBget_requiredperformance();
 int	DBget_proxy_lastaccess(const char *hostname, int *lastaccess, char **error);
 
 char	*DBdyn_escape_string(const char *src);
-char	*DBdyn_escape_string_len(const char *src, int max_src_len);
+char	*DBdyn_escape_string_len(const char *src, size_t max_src_len);
 
 #define ZBX_SQL_LIKE_ESCAPE_CHAR '!'
 char	*DBdyn_escape_like_pattern(const char *src);
